SOURCE/requirements.txt
# VPBank StreamGuard - Fraud Detection System
# Requirements for complete system

# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Machine Learning
xgboost>=2.0.0
lightgbm>=4.0.0

# Deep Learning (for Autoencoder)
tensorflow>=2.13.0

# Explainability
shap>=0.42.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Demo & UI
streamlit>=1.28.0

# AWS SDK
boto3>=1.28.0
botocore>=1.31.0

# Utilities
pyyaml>=6.0
python-dateutil>=2.8.0
pytz>=2023.3
joblib>=1.3.0

# Monitoring
cloudwatch>=1.0.0

# Optional: Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
